WIP: refactor: moved spell checker to services
