+++ pydatacuration/checker/misc_checker.py
@@ -4,8 +4,10 @@
 from loguru import logger
 
 from pydatacuration.checker.check_result_writer import CheckResultWriter
+from pydatacuration.checker.services.spell_checker import SpellCheckerCustomized
 from pydatacuration.services.api_calls.call_dv import DVAPICalls
 from pydatacuration.utils.search_ds_meta import get_depositor_record
+from pydatacuration.utils.search_ds_meta import get_metadata_cm_field
 from pydatacuration.utils.search_result_utils import get_search_result
 
 
@@ -15,6 +17,7 @@ class MiscChecker:
     def __init__(
         self, ds_metadata: dict, check_result_writer: CheckResultWriter, dv_api_calls_instance: DVAPICalls
     ) -> None:
+        """Initialize the MiscChecker class."""
         self.ds_metadata = ds_metadata
 
         # The check result writer instance to write the check results to the database
@@ -22,6 +25,9 @@ class MiscChecker:
 
         # API calls service
         self.dv_api_calls = dv_api_calls_instance
+
+        # Spell checker instance
+        self.spell_checker = SpellCheckerCustomized()
 
     def check_depositor_record(self, collection_alias: str | None = None) -> None:
         """Check if the depositor has deposited data in the dataverse collection.
@@ -60,3 +66,39 @@ class MiscChecker:
 
         else:
             logger.debug('No valid depositor found from the dataset metadata. Skipping depositor history check.')
+
+    def check_spelling(self) -> None:
+        """Check for spelling mistakes in the metadata."""
+        potential_typos = []
+
+        field_list = ['title', 'subtitle', 'alternativeTitle', 'dsDescription.dsDescriptionValue', 'notesText']
+
+        for field in field_list:
+            return_value, field_exists = get_metadata_cm_field(self.ds_metadata, field)
+
+            if field_exists:
+                typos = self.spell_checker.check_spelling(return_value[0])
+                if typos:
+                    typo_messages = {f'{field}: `{item}`' for item in typos}
+                    for message in typo_messages:
+                        logger.info(f'Spelling mistake found in the {field}: {message}')
+
+                    # Collect typos for new structure
+                    for typo in typos:
+                        potential_typos.append(
+                            {
+                                'field': field,
+                                'typo': typo,
+                                'context': return_value[0][:100] + '...'
+                                if len(return_value[0]) > 100
+                                else return_value[0],
+                            }
+                        )
+
+        self.check_result_writer.write(
+            check_name='Fields for Title, Subtitle, Alternative Title, Description, and Notes',
+            check_id='potential_typos',
+            description='Fields for Title, Subtitle, Alternative Title, Description, and Notes',  # noqa: E501
+            unit='typo',
+            results=potential_typos,
+        )
